works normaly with sphere but need to fix orientation cam causing invisible sphere if they're not aligned with orientation

+++ includes/image.h
@@ -6,7 +6,7 @@
 /*   By: spoliart <spoliart@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/25 23:55:57 by spoliart          #+#    #+#             */
-/*   Updated: 2021/12/24 14:30:50 by spoliart         ###   ########.fr       */
+/*   Updated: 2021/12/30 00:28:06 by spoliart         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 typedef struct s_hit
 {
-	int		dist;
+	double	dist;
 	t_coord	pos;
 	t_color	color;
 }			t_hit;
